% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_bin_size.R
\name{fd_bin_size}
\alias{fd_bin_size}
\title{fd_bin_size}
\usage{
fd_bin_size(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the appropriate bin size according to the FD rule
}
\description{
little helper function to calculate hist bin size using the Freedman Diaconis Rule https://en.wikipedia.org/wiki/Freedman%E2%80%93Diaconis_rule
}

