% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_plot.R
\name{gbm_plot}
\alias{gbm_plot}
\title{gbm_plot}
\usage{
gbm_plot(x)
}
\arguments{
\item{gbm.step.oject}{A gbm.step object to be plotted}
}
\value{
A ggplot2 plot
}
\description{
\code{plot_gbm} plots the fitted values against the real values
}
\details{
This is a development function that aims to improve the default plots used
in gbm.step by elith et al. Their plots are good, but I want them to be
better. In the future, this might even be implemented in ggvis.
NB:
This currently only works for gbm.step functions. In the
future more features will be added so that it works for many
decision trees.
}
\examples{

\\code{gbm.step} object
fit.gbm.step <- gbm.step(data = iris,
                         gbm.x = c(2:5),
                         gbm.y = 1,
                         tree.complexity = 2,
                         family = "gaussian",
                         learning.rate = 0.01,
                         bag.fraction = 0.5)

gbm_plot(fit.gbm.step)

# with piping
fit.gbm.step \%>\% plot_gbm.step

Unfortunately it cannot yet run a \\code{gbm} object:
\dontrun{

gbm.fit <- gbm(Sepal.Width ~ .,
               distribution = "gaussian",
               data = iris)

gbm_plot(gbm.fit)
}

}

