% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_missing.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{ggplot_missing}
\usage{
ggplot_missing(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
A ggplot of the missing data
}
\description{
A function that plots missing data in ggplot
}

