% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importance_plot.R
\name{importance_plot}
\alias{importance_plot}
\title{importance_plot}
\usage{
importance_plot(x)
}
\arguments{
\item{x}{is an rpart or gbm.step object}
}
\value{
a ggplot plot of the variable importance
}
\description{
\code{importance_plot} make a graph of variable importance
}
\details{
takes an `rpart` or `gbm.step` fitted object and makes a plot of variable
importance
}
\examples{
turn a fitted object into a data_frame

 \\code{rpart} object
fit.rpart <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)

importance_plot(fit.rpart)

# you can even use piping

fit.rpart \%>\% importance_plot
}

