% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/importance_table.R
\name{importance_table}
\alias{importance_table}
\title{importance_table}
\usage{
importance_table(x)
}
\arguments{
\item{x}{An rpart of gbm.step object to be turned into a tidy dataframe
    containing importance values

This currently only works for rpart and gbm.step functions
need to add some features}
}
\description{
A function that turns the variable importance of BRT into a dataframe
}

