% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imputation_plot.R
\name{imputation_plot}
\alias{imputation_plot}
\title{imputation_plot}
\usage{
imputation_plot(mice.object, vars, include.data)
}
\arguments{
\item{mice.object}{is an object of class `mids`}

\item{vars}{a vector of variables as character}

\item{include.data}{logical: Do you want to include the original data?}
}
\value{
a density plot from the ggplot2 package
}
\description{
\code{imputation_plot} graphically compare values across imputations
}
\details{
This code takes an object fitted with the `mice` package and plots a density
of each of the specified variables. Currently the variables have to be
specified but in the future they might not be
}
\examples{
imp <- mice(nhanes)
   plot1 <- imputation_plot(mice.object = imp,
                            vars = c("age", "bmi"),
                            include.data = F)

   # show the density
   plot1$density

Required packages

dplyr
reshape2
mice
ggplot2
}

